[Update] 컨트롤러 분리(rest controller / controller) 및 예약 기능 rest controller로 수정
METABUS-28
