+++ src/main/java/com/metanet/metabus/bus/controller/BusController.java
@@ -4,12 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequiredArgsConstructor
 public class BusController {
-
 
     @GetMapping("/bus/timetable")
     public String searchBus(
@@ -25,6 +25,17 @@ public class BusController {
         model.addAttribute("roundTrip", roundTrip);
 
         return "bus/busTimeTable";
+    }
+
+    @GetMapping("/bus/reservation")
+    public String makeReservation() {
+        return "bus/bus-reservation-table";
+    }
+
+    @PostMapping("/bus/payment")
+    public String getPaymentList(@RequestParam("data") Long[] reservationIds, Model model) {
+
+        return "redirect:/";
     }
 
 }
