+++ src/main/java/com/metanet/metabus/bus/controller/BusRestController.java
@@ -2,31 +2,32 @@ package com.metanet.metabus.bus.controller;
 
 import com.metanet.metabus.bus.dto.ReservationInfoRequest;
 import com.metanet.metabus.bus.service.ReservationService;
+import com.metanet.metabus.bus.service.SeatService;
 import com.metanet.metabus.member.dto.MemberDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
+import java.time.LocalDate;
 import java.util.List;
 
-@Controller
+@RestController
 @RequiredArgsConstructor
-public class ReservationController {
+public class BusRestController {
 
+    private final SeatService seatService;
     private final ReservationService reservationService;
+
+    @GetMapping("/read/{busNum}/{departureDate}")
+    public List<Long> read(@PathVariable Long busNum, @PathVariable @DateTimeFormat(pattern = "yyyyMMdd") LocalDate departureDate) {
+        return seatService.read(busNum, departureDate);
+    }
 
     @PostMapping("/bus/reservation")
     public List<Long> makeReservation(HttpSession session, @RequestBody ReservationInfoRequest reservationInfoRequest) {
 
         MemberDto memberDto = (MemberDto) session.getAttribute("loginMember");
         return reservationService.create(memberDto, reservationInfoRequest);
-    }
-
-    @GetMapping("/bus/reservation")
-    public String makeReservation() {
-        return "bus/bus-reservation-table";
     }
 }
